Implement basic support for 'revert'
CascadeMap now holds the cascaded declarations for the user-agent, user,
and author origins separately. The 'revert' keyword is resolved to
the cascaded value for the target origin before being applied to
ComputedStyle. (This can happen recursively up to the kUserAgent origin.

Note that 'revert' will not yet work to/from surrogates (css-logical
properties), nor in @keyframes.

Bug: 579788
Change-Id: I98b40d596867129e486035f9cd9f20e7b7cbe08e