[css-filters] Migrate backdrop-filter-rendering.html test to WPT
Migrate this test out of third_party/blink/web_tests/css3/filters
and into the WPT-specific directory, adding links to the relevant
specs and a test assertion describing its purpose.

Since there was already a similar test in WPT, this CL also renames
that test to adding the -001 suffix to it, so that we can consistently
add this new one as a -002 one. Additionally, this CL also adapts a
bit the content of that previously existing test in order to contain
as much metadata as the one being added now, that is, a link to the
original chromium bug, a meta description, links to the same specs
pointed also from the newly-added -002 test, and a consistent title.

Bug: 1063749
Change-Id: I96781e53d1fa272d094937ff1582baac5b82c1b2